Changed putnbr to work with ssize_t instead of int. That way there's no need to convert to long and can receive bigger numbers.

+++ libft.h
@@ -52,7 +52,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len);
 int		ft_putchar_fd(char c, int fd);
 int		ft_putstr_fd(char const *s, int fd);
 int		ft_putendl_fd(char const *s, int fd);
-int		ft_putnbr_fd(int n, int fd);
+int		ft_putnbr_fd(ssize_t n, int fd);
 
 //Others
